scrapper/requirements.txt
pyasn1===0.4.1
scrapy
pytz==2016.10
pymongo==3.5.1
scrapyd==1.1.0
schedule==0.4.3
requests==2.18.4
scrapy-splash==0.7.2
unidecode
python-dotenv==0.9.1
